# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: GTW IOT API
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
...
securityDefinitions:
  UserSecurity:
    type: apiKey
    in: header
    name: x-api-key
security:
  - UserSecurity: []
paths:
  /asset/getAsset:
    # binds a127 app logic to a route
    x-swagger-router-controller: asset
    get:
      tags:
      - asset
      summary: Retrieve list of assets
      description: Return list of assets by user id
      # used as the method name of the controller
      operationId: getAsset
      parameters:
        - name: userID
          in: query
          description: input user ID
          required: true
          type: string
      responses:
        200:
          description: Success
          schema:
            type: object
            properties: 
              Items:
                type: array
                items:
                   $ref: "#/definitions/AssetModel"
        # responses may fall through to errors
        400:
          description: Invaid Input
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Item not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /asset/getSingleAsset:
    # binds a127 app logic to a route
    x-swagger-router-controller: asset
    get:
      tags:
      - asset
      summary: Retrieve list of assets
      description: Return list of assets by company id
      # used as the method name of the controller
      operationId: getSingleAsset
      parameters:
        - name: AssetID
          in: query
          description: Asset id
          required: true
          type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/AssetModel"
        # responses may fall through to errors
        400:
          description: Invaid Input
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: Item not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /asset/addAsset:
    x-swagger-router-controller: asset
    post:
      tags:
      - asset
      summary: Add a new Asset
      description: Add a new asset
      # used as the method name of the controller
      operationId: addAsset
      parameters:
      - in: body
        name: AssetObject
        description: Asset object thats need to be added into table
        required: true
        schema:
          $ref: '#/definitions/AssetModel'
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /asset/updateAsset:
    x-swagger-router-controller: asset
    put:
      tags:
      - asset
      summary: update an existing Asset
      description: update an existing Asset 
      # used as the method name of the controller
      operationId: updateAsset
      parameters:
      - in: body
        name: body
        description: Asset object thats need to be updated
        required: true
        schema:
          $ref: '#/definitions/AssetModel'
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /asset/deleteAsset:
    x-swagger-router-controller: asset
    delete:
      tags:
      - asset
      summary: delete an existing Asset
      description: delete an existing Asset 
      # used as the method name of the controller
      operationId: deleteAsset
      parameters:
      - name: assetID
        in: query
        description: asset ID
        required: true
        type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  AssetModel:
    type: object
    properties:
      UserID: 
        type: string 
      AssetID:
        type: string
      AddTimeStamp:
        type: integer
      DisplayName:
        type: string
      LastestTimeStamp:
        type: integer
      VerificationCode:
        type: string
...
  SuccessResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string